paarriba4

+++ src/main/java/com/once/facturas/model/ProductoRepository.java
@@ -2,7 +2,7 @@ package com.once.facturas.model;
 
 import org.springframework.data.repository.CrudRepository;
 
-public interface ProductoRepository extends CrudRepository<Producto, Long>{
+public interface ProductoRepository extends CrudRepository<Producto, Integer>{
 
 
 }
